+++ packages/plugin-packages/mashroom-portal/test/backend/controllers/ssr-bootstrap.js
@@ -1,0 +1,9 @@
+
+var bootstrap = function bootstrap(portalAppSetup) {
+    return Promise.resolve({
+        html: '<p>server side rendered html</p>',
+        injectHeadScript: 'alert("foo")'
+    });
+};
+
+exports.default = bootstrap;