+++ packages/plugin-packages/mashroom-portal/test/backend/utils/ssr-bootstrap.js
@@ -3,5 +3,4 @@ var bootstrap = function bootstrap(portalAppSetup) {
     return Promise.resolve('this is a test');
 };
 
-var _default = bootstrap;
-exports.default = _default;
+exports.default = bootstrap;